Fix #2337   Flyout style from Accent to Light is not working
