+++ docs/release-notes/1.5.0.md
@@ -41,6 +41,13 @@
     ![mahapps_newunderline4](https://cloud.githubusercontent.com/assets/658431/24204520/0e6f3cbc-0f19-11e7-8a2b-f40752918a96.gif)
 - New underline types also for `AnimatedTabControl` and `AnimatedSingleRowTabControl` styles and for `MetroAnimatedTabControl` and `MetroAnimatedSingleRowTabControl` [#2905](https://github.com/MahApps/MahApps.Metro/pull/2905)
 - Fixed `ClearTypeHint` and wrong `Margin` for `ComboBox` and `ComboBoxItem` [#2903](https://github.com/MahApps/MahApps.Metro/issues/2903)
+- Fix `ListView` style issues
+    + Use accent brush for highlighted border
+    + Fix SnapsToDevicePixels usage
+    + Fix disabled foreground brush
+- Fix SnapsToDevicePixels usage for `CheckBox`
+- Fix `Flyout` modal usage. All title bar commands are now hidden if a `Flyout` is modal.
+- Fix background and foreground usage of controls in a `Flyout` if user changes theme on the fly. [#2337](https://github.com/MahApps/MahApps.Metro/issues/2337) The `Flyout` has a different background, so we should override some background and foreground brushes and colors.
 
 ## Closed Issues
 
@@ -59,3 +66,4 @@
 - [#2895](https://github.com/MahApps/MahApps.Metro/issues/2895) [RFC] [Enhancement] Proposed TabControlHelper.IsUnderlined Change
 - [#2903](https://github.com/MahApps/MahApps.Metro/issues/2903) 1 pixel space between ComboBoxItems closes ComboBox
 - [#2841](https://github.com/MahApps/MahApps.Metro/issues/2841) Modal flyout issue
+- [#2337](https://github.com/MahApps/MahApps.Metro/issues/2337) Flyout style from Accent to Light is not working
