--- src/MahApps.Metro.Samples/MahApps.Metro.Demo/MahApps.Metro.Demo.Shared/ExampleWindows/CustomFlyout.xaml
@@ -1,65 +1,0 @@
-﻿<controls:Flyout x:Class="MetroDemo.ExampleWindows.CustomFlyout"
-                 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                 xmlns:controls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
-                 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-                 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-                 x:Name="customFlyout"
-                 Width="250"
-                 CloseCommand="{Binding ElementName=customFlyout, Path=CloseCmd}"
-                 IsOpen="{Binding ElementName=customFlyout, Path=FlyoutIsOpen}"
-                 Header="Custom Flyout"
-                 Position="Right"
-                 mc:Ignorable="d"
-                 d:DesignHeight="600"
-                 d:DesignWidth="300">
-
-    <controls:Flyout.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="pack://application:,,,/MahApps.Metro;component/Styles/FlatSlider.xaml" />
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </controls:Flyout.Resources>
-
-    <StackPanel Margin="10"
-                Orientation="Vertical">
-
-        <CheckBox Margin="5"
-                  Content="CanCloseFlyout"
-                  IsChecked="{Binding ElementName=customFlyout, Path=CanCloseFlyout}" />
-        <controls:ToggleSwitch Margin="5"
-                               Style="{DynamicResource MahApps.Metro.Styles.ToggleSwitch.Win10}"
-                               OnLabel="Can Close"
-                               OffLabel="Closing disabled"
-                               IsChecked="{Binding ElementName=customFlyout, Path=CanCloseFlyout}" />
-
-        <Slider Margin="5"
-                Style="{StaticResource FlatSlider}"
-                Value="50" />
-
-        <ListView Height="200"
-                  Margin="0, 10, 0, 0"
-                  BorderThickness="0"
-                  ItemsSource="{Binding ElementName=customFlyout, Path=Artists}"
-                  SelectedIndex="0"
-                  Style="{StaticResource VirtualisedMetroListView}">
-            <ListView.View>
-                <GridView>
-                    <GridViewColumn DisplayMemberBinding="{Binding ArtistId}" Header="ID" />
-                    <GridViewColumn DisplayMemberBinding="{Binding Name}" Header="artist" />
-                </GridView>
-            </ListView.View>
-        </ListView>
-
-        <ListBox Height="200"
-                 Margin="0, 10, 0, 0"
-                 BorderThickness="0"
-                 DisplayMemberPath="Name"
-                 ItemsSource="{Binding ElementName=customFlyout, Path=Artists}"
-                 SelectedIndex="0"
-                 Style="{StaticResource VirtualisedMetroListBox}" />
-
-    </StackPanel>
-
-</controls:Flyout>
