+++ src/MahApps.Metro.Samples/MahApps.Metro.Demo/MahApps.Metro.Demo.Shared/ExampleWindows/ShowcaseFlyout.xaml
@@ -1,0 +1,146 @@
+﻿<controls:Flyout x:Class="MetroDemo.ExampleWindows.ShowcaseFlyout"
+                 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+                 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+                 xmlns:controls="http://metro.mahapps.com/winfx/xaml/controls"
+                 xmlns:metroDemo="clr-namespace:MetroDemo"
+                 mc:Ignorable="d"
+                 d:DesignHeight="300"
+                 d:DesignWidth="400"
+                 d:DataContext="{d:DesignInstance metroDemo:MainWindowViewModel}"
+                 CloseCommand="{Binding CloseCmd}"
+                 Header="Flyout Showcase"
+                 Theme="Dark"
+                 Position="Right"
+                 Width="500"
+                 x:Name="ThisView">
+
+    <controls:Flyout.Resources>
+        <ResourceDictionary>
+            <ResourceDictionary.MergedDictionaries>
+                <ResourceDictionary Source="pack://application:,,,/MahApps.Metro;component/Styles/FlatSlider.xaml" />
+            </ResourceDictionary.MergedDictionaries>
+
+            <ObjectDataProvider x:Key="FlyoutThemeEnumValues"
+                                MethodName="GetValues"
+                                ObjectType="{x:Type controls:FlyoutTheme}">
+                <ObjectDataProvider.MethodParameters>
+                    <x:Type TypeName="controls:FlyoutTheme" />
+                </ObjectDataProvider.MethodParameters>
+            </ObjectDataProvider>
+        </ResourceDictionary>
+    </controls:Flyout.Resources>
+
+    <ScrollViewer>
+        <StackPanel Orientation="Vertical"
+                    Margin="2">
+            <Grid Grid.IsSharedSizeScope="True"
+                  Margin="2">
+
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="Auto"
+                                      SharedSizeGroup="LabelsColumn" />
+                    <ColumnDefinition Width="*" />
+                </Grid.ColumnDefinitions>
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="Auto" />
+                    <RowDefinition Height="Auto" />
+                    <RowDefinition Height="Auto" />
+                    <RowDefinition Height="Auto" />
+                </Grid.RowDefinitions>
+
+                <TextBlock Grid.Column="0"
+                           Grid.Row="0"
+                           Margin="2"
+                           VerticalAlignment="Center"
+                           Text="Flyout Theme" />
+                <ComboBox Grid.Column="1"
+                          Grid.Row="0"
+                          Margin="2"
+                          ItemsSource="{Binding Source={StaticResource FlyoutThemeEnumValues}}"
+                          SelectedItem="{Binding ElementName=ThisView, Path=Theme, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
+
+                <CheckBox Grid.Column="1"
+                          Grid.Row="1"
+                          Margin="2"
+                          IsChecked="{Binding CanCloseFlyout, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
+                          Content="Can close Flyout" />
+                <controls:ToggleSwitch Grid.Column="1"
+                                       Grid.Row="2"
+                                       Margin="2"
+                                       IsChecked="{Binding CanCloseFlyout, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
+                                       OnLabel="Allowed"
+                                       OffLabel="Disabled"
+                                       Header="Flyout closing" />
+                <controls:ToggleSwitch Grid.Column="1"
+                                       Grid.Row="3"
+                                       Margin="2"
+                                       IsChecked="{Binding CanCloseFlyout, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
+                                       Style="{DynamicResource MahApps.Metro.Styles.ToggleSwitch.Win10}"
+                                       OnLabel="Allowed"
+                                       OffLabel="Disabled"
+                                       Header="Flyout closing" />
+
+            </Grid>
+
+            <controls:MetroAnimatedTabControl Height="300"
+                                              Margin="2"
+                                              controls:TabControlHelper.Underlined="TabPanel">
+                <TabItem Header="Artists">
+                    <ListView Margin="2"
+                              BorderThickness="0"
+                              ItemsSource="{Binding Artists, Mode=OneWay}"
+                              SelectedIndex="0"
+                              Style="{StaticResource VirtualisedMetroListView}">
+                        <ListView.View>
+                            <GridView>
+                                <GridViewColumn DisplayMemberBinding="{Binding ArtistId}"
+                                                Header="ID" />
+                                <GridViewColumn DisplayMemberBinding="{Binding Name}"
+                                                Header="artist" />
+                            </GridView>
+                        </ListView.View>
+                    </ListView>
+                </TabItem>
+                <TabItem Header="Albums">
+                    <ListBox Margin="2"
+                             BorderThickness="0"
+                             DisplayMemberPath="Title"
+                             ItemsSource="{Binding Albums, Mode=OneWay}"
+                             SelectedIndex="0"
+                             Style="{StaticResource VirtualisedMetroListBox}" />
+                </TabItem>
+                <TabItem Header="Slider">
+                    <StackPanel Orientation="Vertical">
+                        <Slider x:Name="TheSlider"
+                                Margin="4"
+                                LargeChange="10"
+                                Maximum="100"
+                                Minimum="0"
+                                SmallChange="1"
+                                Value="20"
+                                controls:SliderHelper.EnableMouseWheel="MouseHover" />
+                        <Slider Margin="4"
+                                LargeChange="10"
+                                Maximum="100"
+                                Minimum="0"
+                                SmallChange="1"
+                                Style="{StaticResource FlatSlider}"
+                                Value="{Binding ElementName=TheSlider, Path=Value, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
+                        <TextBlock Margin="2"
+                                   HorizontalAlignment="Center"
+                                   Text="{Binding ElementName=TheSlider, Path=Value, Mode=OneWay, StringFormat={}{0:N0}}"
+                                   FontSize="48" />
+                        <TextBox Margin="2"
+                                 Text="{Binding ElementName=TheSlider, Path=Value, Mode=OneWay, StringFormat='{}Slider value is {0:N0}'}"
+                                 controls:TextBoxHelper.Watermark="TextBox 01" />
+                        <TextBox Margin="2"
+                                 controls:TextBoxHelper.Watermark="TextBox 02" />
+                    </StackPanel>
+                </TabItem>
+            </controls:MetroAnimatedTabControl>
+        </StackPanel>
+    </ScrollViewer>
+</controls:Flyout>
+
