+++ src/MahApps.Metro.Samples/MahApps.Metro.Demo/MahApps.Metro.Demo.Shared/ExampleWindows/ShowcaseFlyout.xaml.cs
@@ -1,0 +1,12 @@
+﻿namespace MetroDemo.ExampleWindows
+{
+    using MahApps.Metro.Controls;
+
+    public sealed partial class ShowcaseFlyout : Flyout
+    {
+        public ShowcaseFlyout()
+        {
+            this.InitializeComponent();
+        }
+    }
+}
