+++ src/MahApps.Metro.Samples/MahApps.Metro.Demo/MahApps.Metro.Demo.Shared/MainWindowViewModel.cs
@@ -183,6 +183,36 @@ namespace MetroDemo
             }
         }
 
+        private bool canCloseFlyout = true;
+
+        public bool CanCloseFlyout
+        {
+            get { return this.canCloseFlyout; }
+            set
+            {
+                if (Equals(value, this.canCloseFlyout))
+                {
+                    return;
+                }
+                this.canCloseFlyout = value;
+                this.RaisePropertyChanged("CanCloseFlyout");
+            }
+        }
+
+        private ICommand closeCmd;
+
+        public ICommand CloseCmd
+        {
+            get
+            {
+                return this.closeCmd ?? (this.closeCmd = new SimpleCommand
+                                                         {
+                                                             CanExecuteDelegate = x => this.CanCloseFlyout,
+                                                             ExecuteDelegate = x => ((Flyout)x).IsOpen = false
+                                                         });
+            }
+        }
+
         private ICommand textBoxButtonCmd;
 
         public ICommand TextBoxButtonCmd
