+++ src/MahApps.Metro/MahApps.Metro/Styles/Controls.TreeView.xaml
@@ -81,7 +81,7 @@
     <converters:TreeViewMarginConverter x:Key="LengthConverter" Length="12" />
 
     <Style x:Key="MetroTreeViewItem" TargetType="{x:Type TreeViewItem}">
-        <Setter Property="Foreground" Value="{DynamicResource BlackBrush}" />
+        <Setter Property="Foreground" Value="{DynamicResource TextBrush}" />
         <Setter Property="Background" Value="{DynamicResource WhiteBrush}" />
         <Setter Property="FocusVisualStyle" Value="{x:Null}" />
         <Setter Property="HorizontalContentAlignment" Value="{Binding HorizontalContentAlignment, RelativeSource={RelativeSource AncestorType={x:Type ItemsControl}}}" />
